+++ termbox/CHANGELOG.md
@@ -3,7 +3,7 @@
 - Add user events to `Event` type
 - Overhaul `Attr`+`Cell`+`Cells`+`Cursor` API; now it's `Cell`+`Color`+`Scene`
 - Rename `set` to `cell`
-- Add names for bright versions of basic colors (e.g. `brightRed`)
+- Add combinator to brighten basic colors (`bright`)
 - Add combinator for miscellaneous colors (`color`)
 - Add combinator for monochrome colors (`gray`)
 - Add `Pos` type